とりあえずlog_wait，動いた？

+++ Core/Inc/device/uart.hpp
@@ -3,5 +3,5 @@
 
 void uart_init();
 void uart_write(const char* str);
-void uart_write(const uint8_t* data, uint16_t len);
+void uart_write(const uint8_t* data, uint32_t len);
 void uart_write_format(const char* fmt, ...);
